requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
deepface==0.0.95
numpy>=2.0.0,<2.3.0
Pillow>=10.0.1
opencv-python>=4.12.0
scikit-learn>=1.7.0
tensorflow>=2.20.0
faiss-cpu